Move nexus_plot_load.cpp, nexus_plot_spe1.cpp and nexus_unit.cpp to catch2 framework
